// src/company/dto/updateCompany.dto.d.ts
export declare class UpdateCompanyDto {
    companyName: string;
    email: string;
    contactPerson: string;
    phone: string;
    industry: string;
    location: string;
    size: string;
    aboutUs: string[];
    contactEmail: string;
    contactPhone: string;
    site: string;
    social: string;
    projects: string[];
    reviews: string[];
}